+++ globals_windows.go
@@ -2,7 +2,11 @@
 
 package main
 
-var p pokenv = pokenv{
-	registry:  realRegistry{},
-	pathcheck: check,
+import "unsafe"
+
+var registry = realRegistry{}
+
+func refreshEnvironment() {
+	SendMessageTimeout(HWND_BROADCAST, WM_SETTINGCHANGE, 0,
+		uintptr(unsafe.Pointer(StringToUTF16Ptr("Environment"))), SMTO_ABORTIFHUNG, 5000)
 }