+++ registry_mock.go
@@ -10,12 +10,12 @@ func init() {
 	mock.env = make(map[string]string)
 }
 
-func (r mockRegistry) SetString(path regPath, valueName string, value string) error {
+func (r mockRegistry) SetString(path regKey, valueName string, value string) error {
 	r.env[valueName] = value
 	return nil
 }
 
-func (r mockRegistry) DeleteValue(path regPath, valueName string) error {
+func (r mockRegistry) DeleteValue(path regKey, valueName string) error {
 	delete(r.env, valueName)
 	return nil
 }